move some alientank funcs to commonlib, edit animationlib a bit
